Move more tests to the faster unit test class
