+++ packages/framework/tests/Unit/FeaturedImageUnitTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Unit;
 
+use Hyde\Framework\Exceptions\FileNotFoundException;
 use Hyde\Framework\Features\Blogging\Models\FeaturedImage;
 use Hyde\Testing\UnitTestCase;
 
@@ -112,6 +113,42 @@ class FeaturedImageUnitTest extends UnitTestCase
         $this->assertEquals('local', (new LocalImage)->getType());
         $this->assertEquals('remote', (new RemoteImage)->getType());
     }
+
+    public function testGetContentLength()
+    {
+        $this->assertEquals(0, (new NullImage)->getContentLength());
+        $this->assertEquals(0, (new FilledImage)->getContentLength());
+    }
+
+    public function testFeaturedImageGetContentLengthWithNoSource()
+    {
+        $this->expectException(FileNotFoundException::class);
+        $this->expectExceptionMessage('Image at _media/foo does not exist');
+
+        $image = new FeaturedImage('_media/foo', ...$this->defaultArguments());
+        $this->assertEquals(0, $image->getContentLength());
+    }
+
+    public function testCanConstructFeaturedImageWithRemoteSource()
+    {
+        $image = new FeaturedImage('http/foo', ...$this->defaultArguments());
+        $this->assertInstanceOf(FeaturedImage::class, $image);
+
+        $this->assertEquals('http/foo', $image->getSource());
+    }
+
+    public function testCanConstructFeaturedImageWithHttps()
+    {
+        $image = new FeaturedImage('https/foo', ...$this->defaultArguments());
+        $this->assertInstanceOf(FeaturedImage::class, $image);
+
+        $this->assertEquals('https/foo', $image->getSource());
+    }
+
+    protected function defaultArguments(): array
+    {
+        return ['alt', 'title', 'author', 'authorUrl', 'copyright', 'license', 'licenseUrl'];
+    }
 }
 
 class LocalImage extends FeaturedImage
@@ -136,6 +173,11 @@ class NullImage extends FeaturedImage
     {
         parent::__construct('source');
     }
+
+    public function getContentLength(): int
+    {
+        return 0;
+    }
 }
 
 class FilledImage extends FeaturedImage
@@ -143,5 +185,10 @@ class FilledImage extends FeaturedImage
     public function __construct()
     {
         parent::__construct('source', 'alt', 'title', 'author', 'authorUrl', 'copyright', 'license', 'licenseUrl');
+    }
+
+    public function getContentLength(): int
+    {
+        return 0;
     }
 }
